feat: add test utility method that typed  authorization
